Kge javascope cleanup

+++ javascope/jScope/LocalDataProvider.java
@@ -360,12 +360,10 @@ public class LocalDataProvider extends MdsDataProvider /* implements DataProvide
             long maxSpecific = jScopeFacade.convertToSpecificTime( (long) max);
             long minSpecific = jScopeFacade.convertToSpecificTime( (long) min);
 
-            long dt = ( (long) maxSpecific - (long) minSpecific) / MAX_PIXELS;
+            long dt = (maxSpecific - minSpecific) / MAX_PIXELS;
             limitsExpr = "JavaSetResampleLimits(" + minSpecific + "UQ," +
                 maxSpecific + "UQ," + dt + "UQ)";
-        }
-        else
-        {
+        } else {
             double dt = (max - min) / MAX_PIXELS;
             limitsExpr = "JavaSetResampleLimits(" + min + "," + max + "," + dt +
                 ")";
@@ -373,6 +371,4 @@ public class LocalDataProvider extends MdsDataProvider /* implements DataProvide
         GetFloatNative(limitsExpr);
     }
     boolean supportsLargeSignals() {return false;} //Subclass LocalDataProvider will return false
-
-
 }
